Change AddAlias to AddAlias(Signature, Block)
